# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1-oas3", 
    "title": "Kmart Backend Service Internal"
  },
  "servers": [
    {
      "url": "http://localhost:3534/api/v1"
    },
    {
      "url": "http://103.186.0.54:3534/api/v1"
    }
  ],
	"security": [
		{
			"bearerAuth": []
		}
	],
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
	"components": {
...
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"paths": {
...
		"/settings/encryptPass": {
			"get": {
				"tags": ["CMS Settings"],
				"summary": "Untuk encrypt kata sandi",
				"parameters": [
					{
						"name": "kata",
						"in": "query",
						"style": "simple",
						"explode": false,
						"schema": {
							"type": "string",
							"example": ""
						}
					}
				],
				"responses": {
					"200": {
						"description": "Berhasil"
					},
					"404": {
						"description": "Gagal atau Unauthorized"
					}
				}
			}
		},
...
	}
}